+++ test/eda/subscription_test.exs
@@ -35,7 +35,7 @@ defmodule EDA.SubscriptionTest do
       assert sub.entitlement_ids == ["e1"]
       assert sub.renewal_sku_ids == ["sku2"]
       assert sub.current_period_end == ~U[2026-10-01 00:00:00Z]
-      assert sub.status == 2
+      assert sub.status == :ending
       assert sub.country == "FR"
     end